src/backend/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.0.4

Flask-Cors~=5.0.0
python-dotenv~=1.0.1